Changes in the structure - moved calculate to new file
